Updated plugin to grails 4
