+++ .gitignore
@@ -1,11 +1,15 @@
-target
-out
-*.iml
-*.iws
-*.ipr
+Thumbs.db
+.DS_Store
+.gradle
+build/
+out/
+target/
 .idea
+*.iml
+*.ipr
+*.iws
+.project
 .settings
 .classpath
 *.log
-.project
 web-app/WEB-INF/tld
