--- application.properties
@@ -1,4 +1,0 @@
-#Grails Metadata file
-#Wed Feb 22 09:26:11 GMT 2012
-app.grails.version=2.1.1
-app.name=grails-ajax-uploader
