+++ gradle.properties
@@ -1,0 +1,4 @@
+grailsVersion=4.0.11
+org.gradle.daemon=true
+org.gradle.parallel=true
+org.gradle.jvmargs=-Dfile.encoding=UTF-8 -Xmx1024M
