+++ gradlew
@@ -1,43 +1,10 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 ##############################################################################
 ##
-##  Grails start up script for UN*X
+##  Gradle start up script for UN*X
 ##
 ##############################################################################
-
-# Add default JVM options here. You can also use JAVA_OPTS and GRAILS_OPTS to pass JVM options to this script.
-DEFAULT_JVM_OPTS='"-XX:+TieredCompilation" "-XX:TieredStopAtLevel=1" "-XX:CICompilerCount=3"'
-
-# Use the maximum available, or set MAX_FD != -1 to use that value.
-MAX_FD="maximum"
-
-warn() {
-  echo "$*"
-}
-
-die() {
-  echo
-  echo "$*"
-  echo
-  exit 1
-}
-
-# OS specific support (must be 'true' or 'false').
-cygwin=false
-msys=false
-darwin=false
-case "$(uname)" in
-CYGWIN*)
-  cygwin=true
-  ;;
-Darwin*)
-  darwin=true
-  ;;
-MINGW*)
-  msys=true
-  ;;
-esac
 
 # Attempt to set APP_HOME
 # Resolve links: $0 may be a link
@@ -57,7 +24,47 @@ cd "$(dirname \"$PRG\")/" >/dev/null
 APP_HOME="$(pwd -P)"
 cd "$SAVED" >/dev/null
 
-JAR_PATH=$APP_HOME/grails-wrapper.jar
+APP_NAME="Gradle"
+APP_BASE_NAME=$(basename "$0")
+
+# Add default JVM options here. You can also use JAVA_OPTS and GRADLE_OPTS to pass JVM options to this script.
+DEFAULT_JVM_OPTS='"-Xmx64m"'
+
+# Use the maximum available, or set MAX_FD != -1 to use that value.
+MAX_FD="maximum"
+
+warn() {
+  echo "$*"
+}
+
+die() {
+  echo
+  echo "$*"
+  echo
+  exit 1
+}
+
+# OS specific support (must be 'true' or 'false').
+cygwin=false
+msys=false
+darwin=false
+nonstop=false
+case "$(uname)" in
+CYGWIN*)
+  cygwin=true
+  ;;
+Darwin*)
+  darwin=true
+  ;;
+MINGW*)
+  msys=true
+  ;;
+NONSTOP*)
+  nonstop=true
+  ;;
+esac
+
+CLASSPATH=$APP_HOME/gradle/wrapper/gradle-wrapper.jar
 
 # Determine the Java command to use to start the JVM.
 if [ -n "$JAVA_HOME" ]; then
@@ -82,7 +89,7 @@ location of your Java installation."
 fi
 
 # Increase the maximum file descriptors if we can.
-if [ "$cygwin" = "false" -a "$darwin" = "false" ]; then
+if [ "$cygwin" = "false" -a "$darwin" = "false" -a "$nonstop" = "false" ]; then
   MAX_FD_LIMIT=$(ulimit -H -n)
   if [ $? -eq 0 ]; then
     if [ "$MAX_FD" = "maximum" -o "$MAX_FD" = "max" ]; then
@@ -97,11 +104,16 @@ if [ "$cygwin" = "false" -a "$darwin" = "false" ]; then
   fi
 fi
 
+# For Darwin, add options to specify how the application appears in the dock
+if $darwin; then
+  GRADLE_OPTS="$GRADLE_OPTS \"-Xdock:name=$APP_NAME\" \"-Xdock:icon=$APP_HOME/media/gradle.icns\""
+fi
+
 # For Cygwin, switch paths to Windows format before running java
 if $cygwin; then
   APP_HOME=$(cygpath --path --mixed "$APP_HOME")
+  CLASSPATH=$(cygpath --path --mixed "$CLASSPATH")
   JAVACMD=$(cygpath --unix "$JAVACMD")
-  JAR_PATH=$(cygpath --path --mixed "$JAR_PATH")
 
   # We build the pattern for arguments to be converted via cygpath
   ROOTDIRSRAW=$(find -L / -maxdepth 1 -mindepth 1 -type d 2>/dev/null)
@@ -112,8 +124,8 @@ if $cygwin; then
   done
   OURCYGPATTERN="(^($ROOTDIRS))"
   # Add a user-defined pattern to the cygpath arguments
-  if [ "$GRAILS_CYGPATTERN" != "" ]; then
-    OURCYGPATTERN="$OURCYGPATTERN|($GRAILS_CYGPATTERN)"
+  if [ "$GRADLE_CYGPATTERN" != "" ]; then
+    OURCYGPATTERN="$OURCYGPATTERN|($GRADLE_CYGPATTERN)"
   fi
   # Now convert the arguments - kludge to limit ourselves to /bin/sh
   i=0
@@ -142,10 +154,19 @@ if $cygwin; then
   esac
 fi
 
-# Split up the JVM_OPTS And GRAILS_OPTS values into an array, following the shell quoting and substitution rules
-function splitJvmOpts() {
-  JVM_OPTS=("$@")
+# Escape application args
+save() {
+  for i; do printf %s\\n "$i" | sed "s/'/'\\\\''/g;1s/^/'/;\$s/\$/' \\\\/"; done
+  echo " "
 }
-eval splitJvmOpts $DEFAULT_JVM_OPTS $JAVA_OPTS $GRAILS_OPTS
+APP_ARGS=$(save "$@")
 
-exec "$JAVACMD" -jar "${JVM_OPTS[@]}" "$JAR_PATH" "$@"
+# Collect all arguments for the java command, following the shell quoting and substitution rules
+eval set -- $DEFAULT_JVM_OPTS $JAVA_OPTS $GRADLE_OPTS "\"-Dorg.gradle.appname=$APP_BASE_NAME\"" -classpath "\"$CLASSPATH\"" org.gradle.wrapper.GradleWrapperMain "$APP_ARGS"
+
+# by default we should be in the correct project dir, but when run from Finder on Mac, the cwd is wrong
+if [ "$(uname)" = "Darwin" ] && [ "$HOME" = "$PWD" ]; then
+  cd "$(dirname "$0")"
+fi
+
+exec "$JAVACMD" "$@"
