+++ grails-app/conf/AjaxUploaderResources.groovy
@@ -1,8 +1,8 @@
-modules = {
-
-    fileuploader {
-        resource url:'/js/fileuploader.js'
-        resource url:'/css/uploader.css'
-    }
-
-}
+//modules = {
+//
+//    fileuploader {
+//        resource url:'/js/fileuploader.js'
+//        resource url:'/css/uploader.css'
+//    }
+//
+//}
