--- grails-app/controllers/uk/co/desirableobjects/ajaxuploader/AjaxUploadController.groovy
@@ -1,54 +1,0 @@
-package uk.co.desirableobjects.ajaxuploader
-
-import grails.converters.JSON
-
-import static grails.util.Holders.config as grailsConfig
-import org.springframework.http.HttpStatus
-import uk.co.desirableobjects.ajaxuploader.exception.FileUploadException
-import org.springframework.web.multipart.MultipartHttpServletRequest
-import org.springframework.web.multipart.commons.CommonsMultipartFile
-import org.springframework.web.multipart.MultipartFile
-import javax.servlet.http.HttpServletRequest
-
-class AjaxUploadController {
-
-    AjaxUploaderService ajaxUploaderService
-
-    def upload = {
-        try {
-
-            File uploaded = createTemporaryFile()
-            InputStream inputStream = selectInputStream(request)
-
-            ajaxUploaderService.upload(inputStream, uploaded)
-
-            return render(text: [success:true] as JSON, contentType:'text/json')
-
-        } catch (FileUploadException e) {
-
-            log.error("Failed to upload file.", e)
-            return render(text: [success:false] as JSON, contentType:'text/json')
-
-        }
-
-    }
-
-    private InputStream selectInputStream(HttpServletRequest request) {
-        if (request instanceof MultipartHttpServletRequest) {
-            MultipartFile uploadedFile = ((MultipartHttpServletRequest) request).getFile('qqfile')
-            return uploadedFile.inputStream
-        }
-        return request.inputStream
-    }
-
-    private File createTemporaryFile() {
-        File uploaded
-        if (grailsConfig.imageUpload?.containsKey('temporaryFile')) {
-            uploaded = new File("${grailsConfig.imageUpload.temporaryFile}")
-        } else {
-            uploaded = File.createTempFile('grails', 'ajaxupload')
-        }
-        return uploaded
-    }
-
-}
