+++ grails-app/controllers/uk/co/desirableobjects/ajaxuploader/UrlMappings.groovy
@@ -1,0 +1,12 @@
+package uk.co.desirableobjects.ajaxuploader
+
+class UrlMappings {
+
+    static mappings = {
+        "/$controller/$action?/$id?(.$format)?" {
+            constraints {
+                // apply constraints here
+            }
+        }
+    }
+}
