+++ grails-app/init/uk/co/desirableobjects/ajaxuploader/Application.groovy
@@ -1,0 +1,13 @@
+package uk.co.desirableobjects.ajaxuploader
+
+
+import grails.boot.GrailsApp
+import grails.boot.config.GrailsAutoConfiguration
+import grails.plugins.metadata.PluginSource
+
+@PluginSource
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
