+++ grails-app/init/uk/co/desirableobjects/ajaxuploader/BootStrap.groovy
@@ -1,0 +1,9 @@
+package uk.co.desirableobjects.ajaxuploader
+
+class BootStrap {
+
+    def init = { servletContext ->
+    }
+    def destroy = {
+    }
+}
