+++ src/main/groovy/uk/co/desirableobjects/ajaxuploader/exception/FileUploadException.groovy
