+++ src/main/groovy/uk/co/desirableobjects/ajaxuploader/exception/InvalidAttributeValueException.groovy
