+++ src/main/groovy/uk/co/desirableobjects/ajaxuploader/exception/MissingRequiredAttributeException.groovy
