+++ src/main/groovy/uk/co/desirableobjects/ajaxuploader/exception/UnknownAttributeException.groovy
